requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1

